+++ test/parse.test.js
@@ -166,7 +166,20 @@ describe('bigNarrowPeak', () => {
     expect(f1).toMatchSnapshot()
   })
 })
+describe('bigNarrowPeak with regularize', () => {
+  let p
+  beforeAll(() => {
+    p = new BED({ type: 'bigNarrowPeak' })
+  })
 
+  it('bigNarrowPeak', () => {
+    const f1 = p.parseLine(
+      'chr1\t566753\t566953\t.\t468\t.\t103.84\t5.54347\t4.80079\t154',
+      { regularize: true },
+    )
+    expect(f1).toMatchSnapshot()
+  })
+})
 test('real world', () => {
   const autoSql = `table hg18KGchr7
     "UCSC Genes for chr7 with color plus GeneSymbol and SwissProtID"
@@ -189,9 +202,15 @@ test('real world', () => {
 
 test('generate README output', () => {
   const p = new BED()
-  console.log(p.parseLine('chr1\t0\t100'))
-  const p2 = new BED({ type: 'bigGenePred' })
-  const line =
-    'chr1\t11868\t14409\tENST00000456328.2\t1000\t+\t11868\t11868\t255,128,0\t3\t359,109,1189,\t0,744,1352,\tDDX11L1\tnone\tnone\t-1,-1,-1,\tnone\tENST00000456328.2\tDDX11L1\tnone'
-  console.log(p2.parseLine(line))
+  const f1 = p.parseLine('chr1\t0\t100')
+  // outputs { chrom: 'chr1', chromStart: 0, chromEnd: 100 }
+  const f2 = p.parseLine('chr1\t0\t100', { regularize: true, uniqueId: 1 })
+  // outputs { uniqueId: 1, refName: 'chr1', start: 0, end: 100 }
+  console.log(f1)
+  console.log(f2)
+  // console.log(p.parseLine('chr1\t0\t100'))
+  // const p2 = new BED({ type: 'bigGenePred' })
+  // const line =
+  //   'chr1\t11868\t14409\tENST00000456328.2\t1000\t+\t11868\t11868\t255,128,0\t3\t359,109,1189,\t0,744,1352,\tDDX11L1\tnone\tnone\t-1,-1,-1,\tnone\tENST00000456328.2\tDDX11L1\tnone'
+  // console.log(p2.parseLine(line))
 })